Add custom page template for Pickup Games.
This is a one-off single page template designed specifically for the
Pickup Games page.  It has the Google map featuring the pickup game
locations on Oahu hardcoded as the top sidebar widget.
The widgets that are placed into the "Right Sidebar Half Page" will
appear, but they will appear below the Google map.